+++ .gitignore
@@ -1,7 +1,7 @@
 .DS_Store
 node_modules
 /dist
-/tests/unit/coverage/*
+/tests/feature/coverage/*
 /docs
 
 # local env files
