+++ tests/feature/specs/props/axis.spec.js
@@ -4,44 +4,46 @@ import syn from 'syn'
 
 let wrapper
 
-describe('aspect ratio', function () {
-  it('should provide the lockAspectRatio as prop', function () {
-    wrapper = mount(VueDraggableResizable, {
-      propsData: {
-        lockAspectRatio: true
-      }
-    })
-
-    expect(wrapper.props().lockAspectRatio).to.be.true
-  })
-
-  it('should resize the component accordingly to its aspect ratio if lockAspectRatio is true', function (done) {
+describe('axis prop', function () {
+  it('should provide the draggable `axis` prop to the component', function () {
     wrapper = mount(VueDraggableResizable, {
       attachToDocument: true,
       propsData: {
-        w: 200,
-        h: 100,
-        active: true,
-        lockAspectRatio: true
+        axis: 'x'
+      }
+    })
+
+    expect(wrapper.props().axis).to.equal('x')
+  })
+
+  it('should effectively drag the component only on the provided axis', function (done) {
+    wrapper = mount(VueDraggableResizable, {
+      attachToDocument: true,
+      propsData: {
+        axis: 'x',
+        x: 0,
+        y: 0,
+        w: 100,
+        h: 100
       }
     })
 
     wrapper.vm.$nextTick(() => {
       const $el = wrapper.vm.$el
 
-      const rect = $el.querySelector('div.handle-mr').getBoundingClientRect()
+      const rect = $el.getBoundingClientRect()
       const fromX = rect.left
       const fromY = rect.top
 
       syn.drag(
-        $el.querySelector('div.handle-mr'),
+        $el,
         {
           from: { pageX: fromX, pageY: fromY },
-          to: { pageX: fromX + 100, pageY: fromY }
+          to: { pageX: fromX + 100, pageY: fromY + 20 },
+          duration: 10
         },
         function () {
-          expect($el.style.width).to.equal('300px')
-          expect($el.style.height).to.equal('150px')
+          expect($el.style.transform).to.equal('translate(100px, 0px)')
 
           done()
         }
